1st step refactor
